tools/metrics/requirements.txt
python-docx>=1.1.0
pymupdf>=1.24.0
uharfbuzz>=0.39.0
fonttools>=4.50.0
